Add failover mode
Standby mode means that in the case where there are N elastic ips
available, the Nth+1 instance will put itself into standby mode,
periodically checking for an available EIP and issuing a critical
alert until it finds one. However, there are cases where the extra
instance is acting as a failover, so that when one of the instances
with an EIP fails, it will come out of standby and automatically
take over from the failed instance. When this is the goal, we dont
want the standby instance to issue critical alerts as it is the
desired behaviour. This change simply add the --enable-failover-mode
option to the autoeips script to enable this behaviour.